requirements.txt
# Core AG2 and Vertex AI
ag2>=0.9.1
ag2[gemini]>=0.9.1
google-cloud-aiplatform>=1.90.0
python-dotenv>=1.0.0

# Web and Network Tools
dnspython>=2.7.0
python-whois>=0.9.5
requests>=2.32.0
beautifulsoup4>=4.13.0
google-api-python-client>=2.169.0
googlesearch-python>=1.3.0
colorama>=0.4.6
termcolor>=3.1.0
httpx>=0.28.0
httpx-sse>=0.4.0

# SSL Analysis
ssl>=1.16
cryptography>=42.0.0

# CORS Checking
aiohttp>=3.11.0

# WAF Detection
requests-toolbelt>=1.0.0

# CMS Detection
lxml>=4.9.0
urllib3>=2.4.0

# Port Scanner
# No additional requirements (uses built-in socket)

# Endpoint Crawler
httpx>=0.28.0 # Already specified above
playwright>=1.52.0

# Testing
pytest>=8.3.0
pytest-asyncio>=0.26.0

# Concurrency
futures>=3.0.0; python_version < '3.2'